Fix NPE in TJSON encoder
Summary: TJSON encoder was referencing null data structure when trying to read escaped unicode hex characters. This results in a NPE and causes deserialization to fail. This diff fixes this to ensure that bytes are read directly from the underlying ByteBuf rather than the “transport” which will always be null for the ByteBuf based decoders.

Differential Revision: D74127572

fbshipit-source-id: ebcb1d7285ab09e164f80c2f57a0a5c218df6a59