Call exception methods on ContextStack to bump fb303 counters
Summary:
Context: https://fb.workplace.com/groups/rust.language/permalink/1774792253450340/

tl;dr;

We aren't bumping the counters for num_exceptions in the thrift ODS stats. They stay statically at 0, and it seems this is done by the ThriftStatsModuleLight layer which is tied to the context stack callback of when an exception has occurred.

In the stat handler, specifically calling `handlerError` is what bumps this counter: https://fburl.com/code/ug6acac9

There's probably other service framework module impact here, but we'll start with this. There's a lot on the context stack that might be missing, but given this is the first complaint, that's probably OK.

Reviewed By: Imxset21

Differential Revision: D74402604

fbshipit-source-id: a2e0ec18881a28e2b34332902e6c77a90dc3d98b