Fix a deadlock in ThreadManagerExecutorAdapter on libc++
Summary:
libc++ and libstdc++ disagree on the order of destruction of `std::vector` elements. libstdc++ destroys them from the first to the last while libc++ destroys them in the reverse order. This results in a deadlock in the `ThreadManagerExecutorAdapter` dtor on libc++ because there are some dependencies between executors stored in the `owning_` vector. Fix it by explicitly destroying executors in the forward order.

This fixes server tests that use `ThreadManagerExecutorAdapter` which is 6 out of 9 remaining failures, namely:
* `ThreadManagerAdapterOverMeteredExecutorUpstreamPriorities/*`
* `ThreadManagerAdapterSinglePool/*`
* `ThreadManagerAdapterManyPools/*`

Reviewed By: yfeldblum, ot

Differential Revision: D74437306

fbshipit-source-id: fdbebb7060e8fdebfd054a056f08472d74d18836