fix skipped client tests
Summary:
These `async` tests fail, but they're accidentally skipped instead because the unit test has the wrong base class for `async` cases. Fixing that reveals 4 broken tests (in py3) and an additional broken one in thrift-python. A couple expected differences revealed:
- thrift-python clients accept `int` for `enum` behaviors. This is expected after combining thrift-py3 enum with thrift-python enum (latter has int base)
- thrift-python clients accept `None` arguments. That's because they transform the arguments into a thrift-python `Struct`, which inherently accepts unset `None` field values in constructor.

Meanwhile, to make sure we have the desired coverage, taking the calls that were intended to succeed and wrapping them in `assertRaises(TransportError)`. These fail because this handler isn't setup in a way that support connection. Contrast with `client_server.py` test, or the `python/client/test/{a,}sync_client_test.py` tests, where the handler goes in a different process or event loop.

Reviewed By: yoney

Differential Revision: D74774866

fbshipit-source-id: bbec74d8badc7a0b8c67660d4ce7893e98696ca6