Store whole schema bundles in SchemaRegistry
Summary: Schema data is much easier to consume given an invariant that any definition in a bundled schema will have all of its transitive dependencies also present in that schema. To facilitate this we store a pointer to the schema bundle instead of just the root file's schema in the registry.

Reviewed By: praihan

Differential Revision: D75025836

fbshipit-source-id: 604b279ba6b540bf432e88e0f4635ec22652f54b