Fix server tests
Summary:
`ResponseWriteTimeout/*` tests are broken on macOS because they assume that passing zero to `setSendBufSize` will set the buffer size to the minimum value which is only true for Linux. On macOS `setSendBufSize` actually fails with zero. Fix it and check the result of `setSendBufSize`.

This fixes two of the four remaining server test failures.

Do a minor cleanup while at it:

* Rename `TestInteface` to `TestHandler` since it's a service handler.
* Remove unnecessary ctors.
* Use correct types in `sync_echoInt`.
* Make more test targets follow naming conventions.
* Remove unused includes.
* Simplify initialization.

Reviewed By: islamismailov

Differential Revision: D75145517

fbshipit-source-id: 09afbdb290931004a4551549291bbdd10adfad12