mangle exception field names in generated response struct
Summary:
We're relying on user input (`throws` "field" names) for internal variable names. Specifically, we're using these for the generated per-function response `struct`. There's a couple ways for this to be buggy:
1. If an exception is called `"success"`, then it will be impossible to access a non-exception return value.
2. If two exception fields have the same name (as in recent user post), then it will be impossible to access the first error if returned.

The solution is to mangle the internal field name with the field index to ensure there's no chance of collision (leveraging exsting IDL validation). Can't use field key (id) because it may be negative

Changing field names is wire compatible, so it should be ok to push this without explicit compatibility code.

Reviewed By: prakashgayasen

Differential Revision: D75027172

fbshipit-source-id: 67d9f081ceb90f2810a5006b70c92ef04ad5eacd