OrderableTypeUtils: readability improvements
Summary:
Facebook Internal:
This is a re-do of D75105968 (which was reverted by D75201094), with the underlying bug fixed. The issue was in the refactor of `back_propagate_unorderable_types()`: in my initial attempt at moving froma while loop to a for loop, I missed the fact that we actually do need to get a new iterator on every iteration, because new items may be added during the iteration. In this new version, the while loop is left as-is, and a note is added to prevent future over-eager maintainers from making the same mistake.

Exact difference between this diff (V1) and previous attempt (D75105968): P1819463964

Reviewed By: thedavekwon

Differential Revision: D75231981

fbshipit-source-id: 1f20d699344c841c7b852d725407bedb8d2b9348