SyntaxGraph::asTypeSystem
Summary:
Provides access to SyntaxGraph nodes as TypeSystem nodes.
This approach avoids the need to duplicate the type resolution and incremental bundled data acess in TypeSystem while allowing all runtime operations to be defined solely in terms of the latter.
There is some data duplication due to the parallel node hierarchies, but the copies are performed lazily as data is accessed and then cached.
The basic workflow introduced in this diff is to convert the SyntaxGraph at the top-level to a TypeSystem and then look up nodes by URI there. Future diffs will allow converting individual nodes without taking this trip, and handle nodes without URIs.

Reviewed By: pranavtbhat

Differential Revision: D75173383

fbshipit-source-id: fafdcd767271f358b9d5f3eb4ff33c4144b37bd7