refactor in prep for invoking request interceptors
Summary:
This is a refactor the `dispatchRequest`  method. The newly created sub-methods will each get more logic in the next diff adding service interceptor onRequest invocations.

As a side effect of switching to `IntrusiveSharedPtr`, the `onResponse` interceptors are now invoked, if any are register. The `IntrusiveSharedPtr` instance owned by the `EventCallbackBase` helps avoid lifetime issues in next diff.

Note that this diff doesn't start invoking any service interceptors in thrift-python servers because none are registered with `ServiceFramework`.

Reviewed By: createdbysk

Differential Revision: D75494960

fbshipit-source-id: ea3cba2650428861b9f0e6a3cf926bc01de3cef3