invoke onRequest interceptors
Summary:
This diff implements `onRequest` interceptor invocation. At this point we're not making the interceptor arguments available to `onRequest` interceptor. However, it shouldn't be too difficult to make `unique_ptr<IOBuf>` available to cpp if someone wants to manually deserialize it. The tricky part is ensuring lifetime without copy. By extension, it wouldn't be difficult to expose to Python if we added some gencode methods to return the arg class (in response to method name string).

As a side effect of switching to `IntrusiveSharedPtr`, the `onResponse` interceptors are already invoked. The `IntrusiveSharedPtr` instance owned by the `EventCallbackBase` also helps avoid lifetime issues when the actual service handler method completes faster than `onRequest`.

Reviewed By: sazonovkirill

Differential Revision: D75494959

fbshipit-source-id: 37148632acb2703f7dbeff8bfb8d95ddde24f746