Implement tuple-destructuring for each blocks
Summary:
This diff implements "tuple destructuring" as specified in the updated spec (D76056600).

We also add `array.enumerate` (similar to Python's `enumerate`) to retain the existing behavior. The rationale for this generalization is so we can support patterns like `first?` and `last?` without adding them directly to the language. See {D76054449}.

Reviewed By: iahs

Differential Revision: D76006180

fbshipit-source-id: cf57f143508415b975e5ca928fa0aa419b5ebe6d