Extend each blocks to support destructuring
Summary:
This feature is similar to how Python is able to destructure tuples in `for` loops.

The current spec only exposes the `index` but in many cases template code care about `first?` and `last?` instead. This mechanism allows us to support these without adding them directly to the language (see `array.enumerate` in stacked diffs):
- {D76056600}
- {D76054449}

Reviewed By: iahs

Differential Revision: D76056600

fbshipit-source-id: 74b22e451a7980c2cc4922a4ececafa451d032bd