Bundle schema data for thrift/annotation and thrift/lib/thrift
Summary:
Several types in these files, including `any.Any` and `thrift.ExceptionMessage`, are useful at runtime.
In order to resolve circular dependency issues between thrift2ast and the Thrift files in these directories, regular schema bundling is disabled for them so the bootstrapping compiler can be used.
Instead of further complicating this bootstrapping mechanism, dump the combined schema for these files into a new header and load it when available.

Reviewed By: pranavtbhat

Differential Revision: D75995382

fbshipit-source-id: 7c419e9421b6eab91dcbb4a0421aedbc7d5ca5bd