optimize encodeThriftFrameworkMetadata
Summary: ThriftFrameworkMetadataUtils::encodeThriftFrameworkMetadata spent [$683.47K gCPU](https://www.internalfb.com/intern/perf/?name=ThriftFrameworkMetadataUtils%3A%3AencodeThriftFrameworkMetadata&type=function&is_regex=0) a year. 50% of the gCPU comes from calling ThriftFrameworkMetadata::getInstanceKey. This is because ThriftFrameworkMetadata $tfm is a thrift struct and we use it as param key. When we  memoize a param key as a thrift struct, it need to first serialize it by calling getInstanceKey. We could use isitfaster to experiement whether we save gCPU after removing the memoize option.

Reviewed By: armanoz

Differential Revision: D76298810

fbshipit-source-id: b850584c300f7452de30d0876baa06054d478f46