Redo "Update some some core thrift logic to produce class pointers"
Summary:
This is mildly jank. There is a place in the codegen that really uses classnames instead of classpointers because it uses and unsafe cast to lie to hack that a class that doesn't yet exist DOES exist. Luckily, because of that, we can be quite confident that we're not actually invoking off that non-existent class.

As such, forward the UNSAFE_CAST to coerce the type to a class pointer to match the schema during codegen.

If anyone breaks this, they should pretty immediately get build errors, so it's not like this could ever actually break prod.

Reviewed By: vassilmladenov

Differential Revision: D76512313

fbshipit-source-id: bf974fa008b26cd2139dc89affe1ee5ae2110135