S532283 Detection - Log clobber of `setConcurrencyLimit`/`setExecutionRate` to thrift_server_events Scuba.
Summary:
Calling `setConcurrencyLimit` or `setExecutionRate` before server startup would appear to have no effect due to a race condition in the `setMaxRequestsCallbackHandle` and `setMaxQpsCallbackHandle`. When calling `setConcurrencyLimit` before server startup, we would properly cancel the `setMaxRequestsCallbackHandle`. However, it would still execute once after the server started up, and it always executes after the `setConcurrencyLimitCallbackHandle`.

The two offending callback handles now check if they were cancelled. If so, we record the event to scuba. This will help us get a better picture of which services to investigate and ensure fixing this behavior won't have a negative effect.

The bug is fixed later in this stack.

Reviewed By: sazonovkirill

Differential Revision: D76703231

fbshipit-source-id: 47f998668868763bdbff881ddf30bfbec5d03c01