Only import programs that are needed in runtime
Summary:
Currently we import all includes even if they are not referenced later.

However, in BuildGraph API, we won't generate runtime dependencies for modules that are not referenced later since it's unnecessary.

This discrepancy breaks Configerator since they won't materialize modules not in BuildGraph dependencies.

This diff changed the logic and removed unnecessary imports in Patch.

Basically it iterates all types in the root program to see whether it referenced programs outside root, and only imports those includes.

The same idea should work for thrift-python imports which I am experimenting in the next stacked diff

Reviewed By: pranavtbhat

Differential Revision: D77348074

fbshipit-source-id: 9686fd975734b2cbeea8ddadb9b5c2422f76169c