make imports more robust to partial rollout
Summary:
A large number of `.thrift` files pick-up dependencies on `fb303.thrift` or some other base class via `service MyService extends fb303.FacebookService {...}`. That makes it hard to do a partial in-place migration based on all these build graph edges to these core service `.thrift` files.

However, this dependency is never used in types, i.e., as a thrift field, so it's not actually a blocker for in-place migration.

To avoid import error from un-migrated thrift file, switch the dependency imports to `.types`, which contains aliases to the `inplace.py` types (or regular cython types if un-migrated).

Reviewed By: prakashgayasen

Differential Revision: D77401568

fbshipit-source-id: 78a400441137506e442f38a02a619219682d4c96