Reduce cost of LargeSendTest
Summary: D77458531 caused the serializer to respect the growth policy for the IOBufQueue even when copying an externally managed buffer. This test serializes a large externally managed buffer, which requires many allocations to serialize and causes the test to time out in dev mode. Copy into a large managed buffer ourselves to optimize this.

Reviewed By: thedavekwon

Differential Revision: D77548252

fbshipit-source-id: 8bd72b80f4ff1cfef77a7a4a586d24f76f5bb329